+++ frontend/config/mediacms.config.templates.js
@@ -32,7 +32,6 @@ module.exports = {
     errorPage: compileStaticTmpl('errorPage.html'),
     aboutPage: compileStaticTmpl('aboutPage.html'),
     termsPage: compileStaticTmpl('termsPage.html'),
-    contactPage: compileStaticTmpl('contactPage.html'),
     signinPage: compileStaticTmpl('signinPage.html'),
     signoutPage: compileStaticTmpl('signoutPage.html'),
     registerPage: compileStaticTmpl('registerPage.html'),