+++ frontend/src/static/js/components/profile-page/ProfilePagesContent.js
@@ -3,7 +3,7 @@ import React from 'react';
 export default class ProfilePagesContent extends React.PureComponent {
   render() {
     return this.props.children ? (
-      <div className={'profile-page-content' + (this.props.enabledContactForm ? ' with-cform' : '')}>
+      <div className="profile-page-content">
         {this.props.children}
       </div>
     ) : null;