+++ frontend/src/templates/config/installation/contents.config.js
@@ -14,11 +14,6 @@ module.exports = {
         link: './terms.html',
         icon: 'description',
       },
-      {
-        text: 'Contact',
-        link: './contact.html',
-        icon: 'alternate_email',
-      },
     ],
     belowNavMenu: null,
     footer: 'Powered by <a href="//demo.mediacms.io" title="mediacms.io" target="_blank">mediacms.io</a>',