+++ users/models.py
@@ -35,7 +35,6 @@ class User(AbstractUser):
     location = models.CharField("Location", max_length=250, blank=True)
     is_editor = models.BooleanField("MediaCMS Editor", default=False, db_index=True)
     is_manager = models.BooleanField("MediaCMS Manager", default=False, db_index=True)
-    allow_contact = models.BooleanField("Whether allow contact will be shown on profile page", default=False)
 
     class Meta:
         ordering = ["-date_added", "name"]